requirements.txt
flask
flask-cors
torch
torchvision
pillow
transformers
pandas
timm
tqdm
